extracted theme generation logic
